Moved usfm package to new 'parse' package and created empty IParser interface
