+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/parse/IParser.kt
@@ -1,0 +1,4 @@
+package org.wycliffeassociates.otter.common.domain.parse
+
+interface IParser {
+}
