+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/parse/usfm/ParseUsfm.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.domain.usfm
+package org.wycliffeassociates.otter.common.domain.parse.usfm
 
 import java.io.File
